backend_new/requirements.txt
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
fastapi
uvicorn
pymupdf
python-jose==3.3.0
passlib==1.7.4
ecdsa==0.19.0
rsa==4.9
# pillow==9.2.0
pytesseract==0.3.8
setuptools>=65.5.1
streamlit
python-dotenv
ollama
chromadb
pdfplumber 
langchain
langchain-ollama
unstructured
unstructured[all-docs]


docling 
PyPDF2 
openai 
langchain 
faiss-cpu 
tiktoken 
langchain-community 
langchain-openai

